+++ app/discover/loading.tsx
@@ -1,6 +1,7 @@
-import React from 'react';
+import React from "react";
+import { Skeleton } from "@/components/ui/skeleton";
 
 export default function Loading() {
   // Add a custom loading Skeleton component
-  return <p>Loading Page</p>;
+  return <Skeleton className="w-full h-full" />;
 }
