+++ app/profile/page.tsx
@@ -2,6 +2,7 @@
 import { useEffect } from "react";
 import { useUser } from "@account-kit/react";
 import { useRouter } from "next/navigation";
+import { Skeleton } from "@/components/ui/skeleton";
 
 function ProfileRedirect() {
   const router = useRouter();
@@ -13,7 +14,7 @@ function ProfileRedirect() {
 
   return (
     <div className="flex items-center justify-center h-screen text-lg font-medium">
-      Loading profile...
+      <Skeleton className="w-full h-full" />
     </div>
   );
 }
