+++ components/home-page/NonLoggedInView.tsx
@@ -3,7 +3,6 @@
 import React from "react";
 import HeroSection from "./HeroSection";
 import { TopChart } from "./TopChart";
-// import FeaturedVideo from './Featured';
 import { TopVideos } from "./TopVideos";
 
 const NonLoggedInView: React.FC = () => {
@@ -11,7 +10,6 @@ const NonLoggedInView: React.FC = () => {
     <div>
       <HeroSection />
       <TopChart />
-      {/* <FeaturedVideo /> */}
       <TopVideos />
     </div>
   );
